sparse_transformer/example/requirement.txt
datasets==2.9.0
pytorch-lightning==1.6.5
pytorch==2.0.0
xformers==0.0.18
timm==0.6.12
transformers==4.25.1
warmup-scheduler==0.3.0